Updated Unit Movement
Majority of the Changers are in UnitScript. Units now rotate and move correctly to target. The target system is halfway finished, the Unit state is no longer being used and unit action will be decided when it reaches the target. A unit will idle when it has no assigned target.

Known problems:
-Units sit at the center of a territory when spawned, and are not assigned team color
-Units will currently not attack enemy unit
-Units do not have proximity attacking implemented yet